requirements.txt
asgiref==3.10.0
Django==5.1.13
django-cors-headers==4.9.0
psycopg2==2.9.11
PyJWT==2.10.1
sqlparse==0.5.3
